agents/coder/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "meta-agent-coder"
version = "0.1.0"
description = "Code generator agent"
requires-python = ">=3.12"
dependencies = [
    "google-adk[a2a]>=0.5.0",
    "meta-agent-shared",
]
